+++ app/models/user.py
@@ -2,7 +2,7 @@
 用戶資料表模型
 """
 
-from sqlalchemy import Column, Integer, String, Float, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, DECIMAL
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.database.connection import Base
@@ -13,7 +13,7 @@ class User(Base):
     
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(255), nullable=False, index=True)
-    cash_balance = Column(Float, default=0.0)
+    cash_balance = Column(DECIMAL(10, 2), default=0.00)
     
     # 建立時間和更新時間
     created_at = Column(DateTime(timezone=True), server_default=func.now())
